Add additional check to turn off NM only when address is assigned to vhost0

+++ deploy/openshift/sources/nm_stop.sh
@@ -2,7 +2,7 @@
 
 while true;
 do
-  if [[ -L "/sys/class/net/vhost0" ]];
+  if [[ -L "/sys/class/net/vhost0" && $(ip address show vhost0 | grep inet[^6]) ]];
   then
           echo "[INFO] Detected vhost0 interface. Stopping NetworkManager..."
           systemctl stop NetworkManager
